Move record_restore_timing into RestoreConfig
