+++ corehq/ex-submodules/casexml/apps/phone/restore.py
@@ -819,3 +819,59 @@ class RestoreConfig(object):
     def delete_cached_payload_if_necessary(self):
         if self.overwrite_cache and self.cache.get(self._restore_cache_key):
             self.cache.delete(self._restore_cache_key)
+
+    def record_timing(self, status):
+        timing = self.timing_context
+        username = self.restore_user.username
+        duration = timing.duration if timing is not None else -1
+        if duration > 20 or status == 412:
+            sync_log = self.restore_state.current_sync_log
+            sync_log_id = sync_log._id if sync_log else None
+            log = logging.getLogger(__name__)
+            log.info("restore %s: user=%s domain=%s status=%s duration=%.3f",
+                     sync_log_id, username, self.domain, status, duration)
+        tags = [
+            u'status_code:{}'.format(status),
+        ]
+        env = settings.SERVER_ENVIRONMENT
+        if (env, self.domain) in settings.RESTORE_TIMING_DOMAINS:
+            tags.append(u'domain:{}'.format(self.domain))
+        if timing is not None:
+            for timer in timing.to_list(exclude_root=True):
+                if timer.name in RESTORE_SEGMENTS:
+                    segment = RESTORE_SEGMENTS[timer.name]
+                    bucket = _get_time_bucket(timer.duration)
+                    datadog_counter(
+                        'commcare.restores.{}'.format(segment),
+                        tags=tags + ['duration:%s' % bucket],
+                    )
+            tags.append('duration:%s' % _get_time_bucket(timing.duration))
+        datadog_counter('commcare.restores.count', tags=tags)
+
+
+RESTORE_SEGMENTS = {
+    "wait_for_task_to_start": "waiting",
+    "FixtureElementProvider": "fixtures",
+    "CasePayloadProvider": "cases",
+}
+
+
+def _get_time_bucket(duration):
+    """Get time bucket for the given duration
+
+    Bucket restore times because datadog's histogram is too limited
+
+    Basically restore frequency is not high enough to have a meaningful
+    time distribution with datadog's 10s aggregation window, especially
+    with tags. More details:
+    https://help.datadoghq.com/hc/en-us/articles/211545826
+    """
+    if duration < 5:
+        return "lt_005s"
+    if duration < 20:
+        return "lt_020s"
+    if duration < 60:
+        return "lt_060s"
+    if duration < 120:
+        return "lt_120s"
+    return "over_120s"
