+++ corehq/ex-submodules/casexml/apps/phone/tasks.py
@@ -5,7 +5,7 @@ from celery.task import periodic_task, task
 from celery.signals import after_task_publish
 from django.conf import settings
 from casexml.apps.phone.cleanliness import set_cleanliness_flags_for_all_domains
-from casexml.apps.phone.utils import delete_sync_logs, record_restore_timing
+from casexml.apps.phone.utils import delete_sync_logs
 
 
 ASYNC_RESTORE_QUEUE = 'async_restore_queue'
@@ -40,7 +40,7 @@ def get_async_restore_payload(restore_config):
     restore_config.timing_context.stop()  # wait_for_task_to_start
     response = restore_config.generate_payload(async_task=current_task)
     restore_config.timing_context.stop()  # root timer
-    record_restore_timing(restore_config, 'async')
+    restore_config.record_timing('async')
 
     # delete the task id from the task, since the payload can now be fetched from the cache
     restore_config.cache.delete(restore_config.async_cache_key)
